requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
click==7.1.2              # via -r requirements.in
docstring-parser==0.7.1   # via -r requirements.in
importlib-metadata==1.6.0  # via inflect
inflect==4.1.0            # via -r requirements.in
jinja2==2.11.2            # via -r requirements.in
markupsafe==1.1.1         # via jinja2
pyyaml==5.3.1             # via -r requirements.in
regex==2020.4.4           # via -r requirements.in
zipp==3.1.0               # via importlib-metadata
